s/assertThrows/assertThrowsJsUnitException/ for tests that are explicitly asserting a JsUnitException type was thrown.
This fix is required for
which makes using assertThrows to catch JsUnitExceptions an error. See that CL
for more context/discussion.

A global presubmit showed your test failing with this change:

The only way this CL could be dangerous is if you weren't intending to catch a
JsUnitException with assertThrows. If you think this is the case, this means
your test has likely been failing in an unexpected way for a while now and will
require triage by the code owner.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=111557666