RELNOTES: Add type information to the SafeHtml#unwrap (and similar) error message.
I got a message "expected object of type SafeHtml, got 'SafeHtml{<blabla>}'" from a failing test. I thought that the SafeHtml object is converted to a string somewhere in the stack and it took me a long time to find this place. It turned out that it really got SafeHtml but it was defined in another file (there were two different definitions of the SafeHtml class). The check for the marker failed for this reason.

It would save me a long time if the type information is in the error message.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=112978914