Allows AbstractDialog to be sealed.
Previously AbstractDialog dynamically added the dialogInternal_ property in a method outside of the constructor, making this class unsuitable to be used with Object.seal. This is especially important if any subclass tries to use goog.defineClass to extend it, as the resulting subclass is sealed by default, and will not work with the previous AbstractDialog implementation.

RELNOTES: n/a
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=113244048