Add nonce support to jsloader scripts and JSONP requests for CSP.
On the jsloader side, the only change necessary is to add 'nonce' to the direct
attribute map so that the attribute is set using element.setAttribute instead
of through element['nonce']. For JSONP, a setNonce method was added to allow
setting a nonce value that the send method then passes to jsloader.

RELNOTES: Added support for CSP nonces in jsloader scripts and JSONP requests.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=114583428