Add a BLOCK_SIZE field to blockcipher, and remove the blocksize parameter from cbc.js and ctr.js.
The blocksize is a property of a blockcipher, both lowering it and increasing it would be weird and certainly dangerous with some modes (like CTR). The few blockciphers with variable blocksize are exceptions not frequently used.

RELNOTES[INC]: Removes the ability to set the blocksize from goog.crypt.cbc and ctr.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=116958180