Check whether the selection is in the editable field before querying for values. Queries return incorrect values if the selection has moved out of the editable field.
This case can happen if you tab in and out of the editable field fast enough.
The reason it happens is because selection updates are on a delay:
https://github.com/google/closure-library/blob/b1c29d2d086eca9a603ab063a6d3957a22b57f25/closure/goog/editor/field.js#L901

RELNOTES: goog.ui.editor.ToolbarController now checks for selection to still be in the field before querying for values when updating.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=121488035