Re-enable webkit testInstanceOf.
The test was broken by the most recent Chrome update that added type inference
for the "name" property. This test is still brittle as it asserts on the exact
error message which, as we found, can change, but this was much easier
and it isn't any worse than it already is.

RELNOTES: na
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=124862700