Update JsTdTestCaseAdapter such that @suppress {duplicate} annotations are actually considered by the compiler.
The existing code structure defines a function and assigns it at the same time, making the @suppress {duplicate} present in the JsDocs of the function a no-op for the compiler. This creates a problem for compiled tests that are coming from jstestdriver since javascript/jstestdriver/externs.js is being pulled in.

RELNOTES: Update JsTdTestCaseAdapter such that @suppress {duplicate} annotations are actually considered by the compiler.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=125513510