Fix initializing of date with year < 100 in datepicker.
RELNOTES: Fix initializing of date with year < 100 in datepicker.
Before the fix: If you init the datepicker with date with year < 100 (eg.23) it will look like it works, but clicking on the grid cell will switch it 1900+ years (eg. 1923).
After the fix: The year < 100 is preserved through the interaction with the control.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=125713669