TL;DR: replaced the '.' with '[\s\S]' in splitRe_ to conform with the regexp from RFC 3986 Appendix B.
------------

The URI parsing RegExp copied from rfc3986 Appendix B was slightly wrong.

The meaning of the '.' character in the RFC is different than the meaning of
the '.' character in JS - specifically, in the RFC the '.' means 'any
character' while in JS it means 'any character other than newlines'.

Because of that - any URI with a newline in the fragment doesn't parse. Note
that a newline anywhere else in the URI parses correctly - the problem is only
in the fragment.

This causes the following description of the goog.uri.utils.split to be
incorrect:

* Never null, since arbitrary strings may still look like path names.

In fact, any string with a newline in the fragment will actually return null!

This is similar to the bug b/28596147 - only in closure (thankfully!) the
functions weren't defined as "will parse anything except #\n". So I hope this
will be much easier to fix...

RELNOTES: goog.Uri no longer fails when there's a newline in the fragment part of the uri.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=129205710