Prepare for goog.dom.TagName type change.
To improve the type precision of created or accessed elements, the type of goog.dom.TagName members is going to change from string to object instance in cl/126080962. This CL prepares for that and adds goog.dom.TagName as an accepted type for the methods that are called with goog.dom.TagName members. Alternatively, it wraps the goog.dom.TagName members into String().

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=129420740