Prepare for goog.dom.TagName type change.
To improve the type precision of created or accessed elements, the type of goog.dom.TagName members is going to change from string to object instance in cl/126080962. This CL prepares for that and changes all {goog.dom.TagName} types to {!goog.dom.TagName}. goog.dom.TagName is currently an enum which is non-nullable but changing it to an object will make it nullable.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=129421640