Assert that goog.dom.safe.setLocationHref is correctly invoked on an object of type Location.
It's important for security that the goog.dom.safe wrapper methods are invoked
on objects of the correct type.  In particular, goog.dom.safe.setLocationHref
should only be used on Location, as it performs the appropriate validation for
that specific context.  If it was used, for instance, on a HTMLLinkElement,
there could be a security bug (because the validation performed is not
sufficient for the context of the href attribute of a <link rel=stylesheet>).

The constraint is enforced using goog.asserts, i.e. only when assertions are
enabled. Furthermore, the assertion accepts values that are not of type Element
or Location at all, to permit the use of mock objects in tests.

RELNOTES[INC]: Assert that goog.dom.safe.setLocationHref is correctly invoked
on an object of type Location.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=129910417