Assert that goog.dom.safe.setImageSrc is correctly invoked on an object of type HTMLImageElement.
It's important for security that the goog.dom.safe wrapper methods are invoked
on objects of the correct type.  In particular, goog.dom.safe.setImageSrc
should only be used on HTMLImageElement, as it performs the appropriate
validation for that specific context.  If it was used on a HTMLScriptElement
however, there could be a security bug (because the validation performed is not
sufficient for <script src>).

The constraint is enforced using goog.asserts, i.e. only when assertions
are enabled. Furthermore, the assertion accepts values that are not of type
Element or Location at all, to permit the use of mock objects in tests.

Tested: Global presubmit.

RELNOTES[INC]: Assert that goog.dom.safe.setImageSrc is correctly invoked on
an object of type HTMLImageElement.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=130233928