Fix bug: incorrect device pixel ratio
Fixes test case for goog.dom.devicePixelRatio and corrects order of
pixel ratio checks.

This fixes the bug that 0.75 will always be the device pixel ratio if
window.devicePixelRatio is not provided by the execution environment.

This bug did not come up earlier because the test-case for this was
incorrect.

Adds mash to the list of authors as stated in the guidelines for
contributing to this repository.