Assert that goog.dom.safe.setAnchorHref is correctly invoked on an object of type HTMLAnchorElement.
It's important for security that the goog.dom.safe wrapper methods are invoked
on objects of the correct type.  In particular, goog.dom.safe.setAnchorHref
should only be used on HTMLAnchorElement, as it performs the appropriate
validation for that specific context.  If it was used on a HTMLLinkElement
however, there could be a security bug (because the validation performed is not
sufficient for <link rel=stylesheet>).

The constraint is enforced using goog.asserts, i.e. only when assertions
are enabled. Furthermore, the assertion accepts values that are not of type
Element or Location at all, to permit the use of mock objects in tests.

Tested: Global presubmit.

RELNOTES[INC]: Assert that goog.dom.safe.setAnchorHref is correctly invoked on
an object of type HTMLAnchorElement.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=130568143