RELNOTES: Use goog.events.KeyHandler in goog.ui.Zippy to correctly handle the keyboard events.
The situation this solves: The header of the zippy contains a checkbox. The checkbox is unable to get the "keypress" event from the space bar. Since the zippy was listening to the "keydown" event and preventDefault was called, the "keypress" event was never fired, thus the checkbox never got it.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=130644595