Automated g4 rollback of changelist 137709055.
*** Reason for rollback ***

Introduces https://b.corp.google.com/issues/32554395

*** Original change description ***

Specify top position when creating dropdown element.

RELNOTES: Specify top position when creating dropdown element. This change is necessary because on some pages, when the element is created and added to the bottom of the page a scrollbar is shown if a border style is present on the element, causing the first call to reposition to return a wrong value. The user will then see the element slightly moved to the left when they type the first letter in the auto complete.

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=137774766