Support ES6 methods when creating a testSuite.
The current logic uses a for...in loop to look for any test properties. This
does not work if the object is an instance of an ES6 class whose methods will
always be non-enumerable.

RELNOTES: Support ES6 methods when creating a testSuite.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=137990795