Make sure Thenable provided by test methods are resolved into a compliant promise.
'then' method of a Thenable is not necessarily safe/fully A+ compliant and may result in hard to debug problems. So in general it is best to 'resolve' provided Thenable to assimilate them into a compliant implementation before using.
https://promisesaplus.com/#point-46

RELNOTES:N/A

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=139821423