Keep trying to fix generate_latest_docs.sh
In particular, this fixes the current error (wrong java version on travis - needs to be Java8) and also fixes a few other problems with the script.  It now runs on all presubmit checks, but only prepares the commit, rather than actually deploying it, for presubmits.  The push has been moved out into a separate push_latest_docs.sh script.