Automated g4 rollback of changelist 140916449.
*** Reason for rollback ***

It seems like this change requires the latest release of the ESF. Until we can guarantee this (which might be a couple of weeks out), we either need to hide this behind a flag or rollback.

*** Original change description ***

Enable client-version support, using the CVER param, which is supported by the existing v8 wire format.

RELNOTES: n/a

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=141360848