Improve the webchannel handshake speed by moving the channel-test to background or to bypass.
The new option is off by default, so this CL should be a strict no-op if the option is not turned on.

The new option should not be turned on against servers built before cl/148391480.

Added "h2" to the client-wire-protocol list to enable concurrent writes. This regression issue doesn't affect chrome, for which we use a private API to detect if SPDY (i.e. http/2 quic too) is enabled. For other browsers, concurrent writes will now be (re-)enabled, e.g. for cloud shell.

Tested with the staging webchannel with http (gsessionid only) and https (gsessionid and cookies).

RELNOTES: new API ("backgroundChannelTest" option in WebChannel)

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=148402633