Check that the xhr_ property is true in getResponseText() and other get*() methods of the testing XhrIo object. This mimics the behavior of the actual XhrIo object, such that calling these methods from outside the callback will return null/undefined/empty.
This is part of a 3 part change.
1) Add the checks that emulate XhrIo's behavior, with the checks disabled by
default.
2) Modify callers in google3 that rely on the current behavior to explicitly
disable the checks.
3) Enable the checks by default.

Note that in order to introduce XhrIo's behavior, simulateResponse() must be
modified to set the xhr/active values to true, since it is expected behavior
that simulateResponse() can be called without ever having called send().

To disable this behavior, for backwards compatibility in tests, a caller may
set goog.testing.net.XhrIo.allowUnsafeAccessToXhrIoOutsideCallbacks to true, though
this is not recommended. The value is temporarily set to true in this change in
order to separate this change from that which sets the value to true in calling
tests.

RELNOTES: Fixes the XhrIo testing library to more closely emulate the actual
one: disabling calls to getResponse*() methods outside of the event callbacks.
The behavior is currently disabled by default, but can be enabled by setting
goog.net.testing.XhrIo.allowUnsafeAccessToXhrIoOutsideCallbacks to false. The
behavior will shortly be made default, and so to maintain the old behavior, one
should set goog.net.testing.XhrIo.allowUnsafeAccessToXhrIoOutsideCallbacks to
true.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=149762952