goog.db.ObjectStore: on add/put, pass key to goog.async.Deferred when resolving success
Currently, when the key is autogenerated by the db, it is not possible to get the key for the add after a successful write.

RELNOTES: In goog.db.ObjectStore.add/put, pass the key to the resolved Deferred on success.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=149788772