Allow specifying row builder in goog.ui.ac.RichRemote.
RELNOTES[NEW]: Currently, goog.ui.ac.RichRemote evals the first element of the response and uses it as a function to pass the remaining elements. This is dangerous (use of eval) and cumbersome (requires exporting the name of the function used in the response). This change allows specifying a row builder which takes both type and the response.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=156161333