Fixes goog.i18n.NumberFormat to actually correctly format small numbers.
Previously, when small numbers and/or high maximum fractional digits were used, it would result in garbage output (non-printable characters) because it was not expecting fracPart to be in scientific notation when it was iterating through its characters and expecting them to be digits. Adds a unit test for this.

RELNOTES: Fixes bug in goog.i18n.NumberFormat that caused small numbers to be formatted incorrectly (with non-printable characters) when a high value for maximum fractional digits was used.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=156317217