Inline createDom attributes.
cl/154420688 adds a rule which bans innerHTML assignment in createDom. If this rule couldn't prove that the attributes don't contain innerHTML then it reports a possible violation. This refactoring changes the code so that this rule can understand it.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=156550976