When display strings and string diffs from assertEquals exceed a threshold length, display them on separate lines.
Currently it's extremely difficult to read test error messages when comparing very long strings, because everything is put on one line.

There are probably more tests to write for this, but I won't bother unless the general concept is approved.

For example, before:

```
Expected <abcdefghijkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkklmnopqrstuvwxyz> (String) but was <abc...xyz> (String)
Difference was at position 3. Expected [...bcdefghijkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkklmnopqrstuvwxy...] vs. actual [...bc...xy...]
```

After:

```
Expected
<abcdefghijkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkklmnopqrstuvwxyz> (String)
but was
<abc...xyz> (String)
Difference was at position 3. Expected
[...bcdefghijkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkklmnopqrstuvwxy...]
vs. actual
[...bc...xy...]
```

Notice how after, the "abc" prefix of the expected and actual are lined up, and the "bc" in the diff are lined up. This makes it *much* easier to visually scan the lines for differences and understand what exactly is different. Furthermore, it can be difficult to even where where expected ends and actual begins with long enough strings (I've spent tens of seconds scanning for the little "but was" string).

RELNOTES: N/A

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=157830816