Make goog.debug.RelativeTimeProvider a proper singleton so that it can be stripped by the compiler when unused.
The singleton is implemented manually. Using goog.addSingletonGetter was attempted in 2014 (cl/60932518 and cl/60947493) but it broke various tests. This might be a compiler bug.

RELNOTES: Make goog.debug.RelativeTimeProvider a proper singleton so that it can be stripped by the compiler when unused.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=160384461