Initialize the test runner using the local, consistent variable.
In jsunit.js we create an instance of goog.testing.TestRunner and store it on global. Then later we pull it off global and initialize it. If something walks on the global, we get confusing results.

TESTED: https://test.corp.google.com/ui#id=OCL:161593025:BASE:161666211:1499957697084:a7790510

RELNOTES:Initialize the test runner using the local, consistent variable.

In jsunit.js we create an instance of goog.testing.TestRunner and store it on global. Then later we pull it off global and initialize it. If something walks on the global, we get confusing results.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=161850335