RELNOTES: Add convenience method to goog.debug that allows for adding context to an error object and simultaneously converting the object to an Error.
This makes the common use case much cleaner. When you catch an error, it is type *, not type Error, so you have to convert it to an error before doing anything with it. Before, we need to do:

catch(e) {
  var err = goog.debug.enhanceError(e);
  goog.debug.errorcontext.addErrorContext(err, 'key', 'value');
  throw err;
}

With this method, it becomes:

catch (e) {
  throw goog.debug.enhanceErrorWithContext(e, {key: 'value'});
}

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=162940155