Automated g4 rollback of changelist 162940155.
*** Reason for rollback ***

broke lots of projects.

*** Original change description ***

RELNOTES: Add convenience method to goog.debug that allows for adding context to an error object and simultaneously converting the object to an Error.

This makes the common use case much cleaner. When you catch an error, it is type *, not type Error, so you have to convert it to an error before doing anything with it. Before, we need to do:

catch(e) {
  var err = goog.debug.enhanceError(e);
  goog.debug.errorcontext.addErrorContext(err, 'key', 'value');
  throw err;
}

With this method, it bec...

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=162972654