Use eval('()') instead of goog.json.unsafeParse in EvalJsonProcessor.
This class is slated for deletion but there's a rollout that needs to happen first. As this is the last user of goog.json.unsafeParse, let's inline the eval so that goog.json.unsafeParse can be deleted.

RELNOTES: n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=164591605