Automated rollback of changelist 162972654.
*** Reason for rollback ***

Roll forward, using goog.provide instead of goog.module for errocontext because goog.module causes an error when used in goog.debug.

*** Original change description ***

Automated rollback of changelist 162940155.

*** Reason for rollback ***

broke lots of tap projects.

*** Original change description ***

RELNOTES: Add convenience method to goog.debug that allows for adding context to an error object and simultaneously converting the object to an Error.

This makes the common use case much cleaner. When you catch an error, it is type *, not type Error, so you have to convert it to an error before doing anything with it. Before, we need to do:

catch(e) {...

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=164637459