RELNOTES[NEW]: Adds the assertion 'assertFinite' to goog.asserts.
'assertNumber' checks that for a value n, typeof n == 'number', but this is also true for NaN and Infinity. This assert is for circumstances where the engineer wishes to verify that for a value n, n is a number but is not NaN, Infinity, or -Infinity.

Motivation: the TypeScript Style Guide explicitly mentions "assertNotNaN", which does not exist. After discussion, this was decided to be a clearer name/behavior.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=168352519