Check the timeout id for null instead of using goog.isNumber().
This makes it play nicer with Zone.js (used in Angular, etc.), which overwrites the setTimeout() et al to return an object instead of a number. That object is then used to keep track of additional data associated with a handle.

RELNOTES: n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=169287032