RELNOTES[NEW]: Add TrustedResourceUrl#cloneWithParams.
goog.net.Jsonp accepts a URL and then adds some parameters to it. I want to change it to accept TrustedResourceUrl instead of string or goog.Uri.

Passing goog.string.Const as URL is not feasible as we don't have a legacy converter for it and there are many call sites we won't be able to convert easily.

Also, the main argument for accepting goog.string.Const in formatWithParams was this: A security reviewer must review all legacy and unchecked conversions used in the code. trustedResourceUrlFromString('https://example.com/') is safe but if it'd be later passed to formatWithParams with {'redirectTo': 'http://evil.com'} then it's not safe anymore. In my opinion, this is a poor argument as the same could be said about formatWithParams(Const.from('https://example.com/'), {'redirectTo': 'http://evil.com'}) which doesn't need to be security reviewed at all (it doesn't use legacy or unchecked conversion).

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=172470757