Rather than retrieve and evaluate the deps/debug loader file write a script tag for them. Also move writing the Closure deps file back to top level.
This is to fix file:// protocol loading.

Note: this required moving the debug loading logic to the bottom of the file so that goog.bind is defined when we create the debug loader (which is now when base is loaded). I'd argue it also adds a bit more organization to the file.

RELNOTES: Fix file:// protocol loading.
Closes #851

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=174733481