Update SliderBase's handling of dragging in RTL mode to correctly set the value based on how much space is remaining to the left of the thumb.
In RTL mode, portion to the right of the slider thumb is considered "on". The original calculation always used the left coordinate of the thumb to determine the value, resulting in erroneous behavior. As an example, in RTL mode, if you put the thumb 25% in from the left, you would expect the value to be 75% of the range but if would incorrectly be calculated as 25%.

The calculation is based on the ratio of (left of thumb / total width). The fix is to use 1 - (left of thumb / total width) in RTL mode.

RELNOTES[INC]: Fix bug in SliderBase's drag handling while in RTL mode to correctly calculate the new slider value.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=177008947