Adds a @record object to the stop callback function in order to pass extra information about the trace being stopped (such as why the trace should be stopped or may be in future when should it be stopped). Currently, it is used when outstanding traces need to be closed (for example to reset the tracer).
Also fixes a small bug: when the trace events thread is cleared (due to resetting or when the events thread is too big) only the already-stopped start events are released and the events/ids of the outstanding traces are kept to avoid creating new start events with duplicated ids.

RELNOTES: fixing a bug to avoid starting traces with the ids of the outstanding traces also adding an optional parameter to the stop callback function.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=182428080