Added goog.testing.assertionfailure withAssertionFailure().
withAssertionFailure will run the code twice to check if it throws the assertionfailure, then run it again to return the value it is supposed to return. It is useful to get the return value when you expect a function to fail assertion.

RELNOTES: Add test utils that allows function to return value after throw.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=182439810