Supports (true) 0-RTT by signaling to the server to upgrade the handshake request directly to a back-channel.
This allows the server to start to deliver messages to the client immediately following the handshake process, which will use a GET with the <fastHandshake> option enabled for QUIC 0-RTT.

This change is supposed to be safe to deploy against old servers which don't yet support this feature. That is, for old servers or for servers that choose to disable the feature by a flag, there should be no behavior change during the handshake.

RELNOTES: support complete 0-RTT handshake with WebChannel.fastHandshake option.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=184595826