Removes functionality to set test execution order via URL parameter.
The URL parameter is read after autoDiscoverTests which is responsible for sorting hence doesn't look like doing anything. I would have tried to fix it but seems like it is better not to have the functionality if nobody is using it.

RELNOTES: n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=184875885