Rework the Closure Debug Loader.
This new model is clearer and allows for more customization. Once we're sure this is final we'll add public documentation with examples on how to customize this. But for now the existing CLOSURE_IMPORT_SCRIPT/CLOSURE_LOAD_FILE_SYNC should still work.

Additionally add some code for ES6 module support, but for now throw an error if ES6 modules are loaded until the code is thoroughly tested and vetted.

RELNOTES: Reworked the Closure Debug Loader to allow for clarity and ES6 module support (latter is behind a flag while in development).

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=189792564