RELNOTES: Report global errors as part of the test results.
In the past we had multiple remote calls from JsTestRunner into the test page. Error handling for the polling and especially race between the remote calls and state changes in the page results in confusing failure modes. A single polling call and a single API for results will be more robust. To get there we need global errors to be part of the results.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=193054840