Automated g4 rollback of changelist 195815180.
*** Reason for rollback ***

RELNOTES: n/a

*** Original change description ***

Add a mapping of keys to keycode and use this in fireNonAsciiKeySequence to populate the keycode field when tests pass in a string as the key. This will result in more complete testing keydown/keypress events, and is necessary to allow tests that make use of fireNonAsciiKeySequence to work correctly with goog.events.KeyHandler, which only uses `keycode` and not the `key` property.

RELNOTES: Attempt to fill in the keyCode value for the events fired by goog.testing.events.fireNonAsciiKeySequence.

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=195822398