Rollforward of cr/196194882: Fix the AvlTree implementation to *be* an AvlTree. It wasn't balancing correctly because it wasn't calculating node heights correctly. I'm not sure if this invalidated the binary search tree properties or not, but at the very least getHeight was incorrect as well as how rigid balancing would be.
Also while we're here upgrade this to a goog.module and make the Node class module scoped.

NEW: Turns out it wasn't goog.module that broke but ES6. So removed the ES6 features.

*** Reason for rollback ***

Rollforward removing ES6 features that broke an internal project.

*** Original change description ***

Automated g4 rollback of changelist 196194882.

*** Reason for rollback ***

Broke internal project that is not goog.module compatible.

*** Original change description ***

Fix the AvlTree implementation to *be* an AvlTree. It wasn't balancing correctly because it wasn't calculating node heights correctly. I'm not sure if this invalidated the binary search tree properties or not, but at the very least getHeight was incorrect as well as how rigid balancing would be.

Also while we're here upgrad...

***

RELNOTES: n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=196907088