Allow calc() and functions used by grid-* CSS properties in SafeStyle.create.
This CL adds the following functions, and modifies the sanitizer to support these functions:

 - calc

calc uses operators +, -, * and / in the function syntax. These operator symbols are added to FUNCTIONS_RE_. They are not added to VALUE_RE_ because the operators do not seem to be used outside of calc.

 - minmax, fit-content, repeat

These are used by grid-* CSS properties to define grid dimensions. It is possible for functions to nest in this context; see testCreate_allowsNestedFunctions() for examples. We allow one level of nesting; further levels of nesting will trigger the "simple functions" assertion.

These functions also use square brackets, [], to delineate ident-tokens. A function hasBalancedSquareBrackets_ is added to ensure that the brackets are matched, there is no nesting or unbalanced bracket, and the content in the brackets is one ident-token. The square bracket symbols [ and ] are added to VALUE_RE and FUNCTION_RE.

Added tests for these new functions.

RELNOTES: Allow calc() and functions used by grid-* CSS properties in SafeStyle.create.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=206172171