Optionally set the focus of the linkdialog to the "text to display" input instead of the url input when the text to display input is empty when the dialog opens.
This should probably be the default behavior, but in order to avoid breaking people who are expecting different behavior this is added as an optional setting that linkdialogpluin users can access by calling focusTextToDisplayOnOpenIfEmpty().

RELNOTES[NEW]: Optionally focuses the "text to display" input instead of the url input when the dialog opens with an empty "text to display" input.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=209864411