Remove goog.module.ModuleManager from CompilerInputWrangler, replacing it with goog.loader.activeModuleManager.
Move the getDefault function in the closure ModuleManager from the first getInstance() call into the load for the file. The CompilerInputWrangler added ModuleManager usages are the first in a typical application, so it is no longer sufficient to wait for the first getInstance() call to provide a getDefault function.

RELNOTES[NEW]: Move activeModuleManager.get() getDefault parameter to a global function.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=209934813