Fully populate rows in goog.ui.PaletteRenderer#setContent.
If this method is called when previous item count is less than grid width, all newly created rows are not fully populated. In this case the first row size was assumed to be the grid width, which is not accurate.

Now we store the grid width as an element attribute during createDom and reuse that value during setContent. This is safe since Palette#setSize is not allowed to be called after createDom.

RELNOTES: Fully populate rows in goog.ui.PaletteRenderer#setContent when previous item count is less than palette grid width.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=214429835