RELNOTES: Fix a loss-of-precision issue in `NumberFormat` that caused some formattings to appear less accurate.
See https://github.com/google/closure-library/issues/916.

This change introduces the `decimalShift_` function and applies it as a primitive throughout the class. This function is expected to be inefficient, but it more accurately describes the intent of the callsites and corrects for possible edge cases related to floating-point numbers. The loss of performance shouldn't be an issue given the intended applications of `NumberFormat`.

Closes #916

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=215245799