Remove the Checkbox#setEnabled override that additionally sets the tabindex on the control element.
Removing the tabindex fix here should be sufficient now since Control#setEnabled calls renderer_.setFocusable.

The default renderer for a Checkbox is a CheckboxRenderer, which doesn't override setFocusable, so it results in ControlRenderer#setFocusable.

which sets the tabindex on the key event target (which by default is the control element, so the logic was just setting it again, but which subclasses may want to override to ensure the tabindex is on some other element).

RELNOTES: Remove the Checkbox#setEnabled since the tabindex fix is already handled by Control#setEnabled calling ControlRenderer#setFocusable on the key event target.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=217571029