Make the type of the Node argument to Document#importNode non-null.
Document#importNode is defined to take a non-null Node argument, and the browser will throw an exception if null is passed in. However, the externs currently specify the type as nullable. This change is required to unblock fixing the externs.

This CL was automatically generated by finding call sites where the type cannot be proven to be non-null, and adding an explicit cast. A better fix may exist (for example, prepending a `!` to the type of an existing variable). If that's the case, please use go/suggestfix to propose an alternative.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=223047656