RELNOTES: Remove an erroneous @this tag from goog.memoize().
goog.memoize() is not (necessarily) called with a `this` context.
Rather, the returned function has a defined `this` type - either the
global object, or, where memoize is used for a class method, the class
instance it is being called on.

This change moves the `@this` tag to the returned function where it
belongs. It was made necessary by an improvement in TypeScript/Clutz
tooling to recognize `this` types.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=224459317