Refactor polymorphism of goog.ui.ComponentUtil to not force goog.events.EventType to be retained wholesale.
EventType is a huge enum with lots of strings. The polymorphism of getMouseEventType() returned the whole enum object itself, forcing compiled code to define the full enum, when only a very small subset of mouse events is needed.

We're seeing a 2kB size increase in an early-load code module to define this giant enum.

This change uses a parallel enum of mouse events when not using pointer events to reduce the scope of the compiled code bloat.

RELNOTES: n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=225052120