Lock down usages of goog.iter as ES6 iterators should be preferred going forward.
This is the same logic as as added in cl/229763933 (which was later rolled back)
but added to the private_conformance_proto, which applies only to Closure code.

The previous CL erronously modified the shared conformance_proto which is also
used by other Closure clients.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=230785332