Removes the aria label of the cell when the content of the cell is removed.
The aria label will be updated when the cell adds a child. This is solving a bug where the aria label was never updated if the content was set more than once on the cell.

RELNOTES: Removes the aria label of palette cell when removing child. The aria label is added back when the child is added to the cell.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=231811203