Accept image/x-icon as a safe mime type.
x-icon is a container for bmp and png which are already whitelisted.
This is the Closure part of cl/232292855 which has to be submitted separately.

RELNOTES[NEW]: image/x-icon is accepted as a safe mime type for object URLs.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=233609863