Fix behavior of DateTime.fromIsoString on dates that have DST boundaries.
The problem can be reproduced with:

  > goog.date.DateTime.fromIsoString("2019-03-10T11:22:33Z").getUTCHours();

  < 10

The reason this happens is simple. When you run DateTime.fromIsoString(), it
first sets the date, then it sets the time. When the time is set, currently, it
handles a couple cases:

  1. There is no timezone specified. The time is simply set using the local time
     Date functions.

  2. There is an offset specified. The specified offset is used to adjust the
     time. The offset is adjusted from the offset of the date.

  3. UTC is specified. The offset is adjusted from the offset of the date.

The first case works as intended, but the latter two are potentially buggy. The
problem lies in the fact that the offset is calculated based on the time at the
beginning of the day, since getTimezoneOffset is called before the time is set
and after the date is set. I do not believe there is a good way to approach this
without special casing localtime, because there is a chicken-and-egg problem.

RELNOTES: Fixed a bug where parsing UTC ISO 8601 date times with DateTime.fromIsoString on dates containing a timezone boundary in the local timezone would sometimes be off by an hour.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=238131068